refactor duplicated plugin setup code
